requirements.txt
langchain-community
langchain-google-genai
faiss-cpu
unstructured
pypdf
python-docx
python-dotenv
streamlit